# rpc: improve `getreceivedby{address,label}` performance