@theStack 2021-12-08
Rebased on master (necessary as the function modified in this PR has moved from `./src/wallet/rpcwallet.cpp` to `./src/wallet/rpc/coins.cpp`).

@shaavan: Thanks for your review! The performance worries on the second commit concerns transactions that are in `wallet.mapWallet` but don't satisfy the `IsMine` predicate (which should be all txs sent from the wallet to an address not owned by the wallet) ; if the search in the output script set (`output_scripts.count(...)`) takes longer than the `IsMine` check, then on master the loop iteration would be faster for those transactions. On the other hand a search in std::set takes O(log n), so it shouldn't be that bad.

Would be very interested in further opinions.