@achow101 2021-12-15
I don't think it's even necessary to use `IsMine`. We already filter for addresses that are ours within the address book itself by setting `CAddressBookData.purpose` to `receive`. So you can just check for that string rather than doing `IsMine`. It will certainly be more efficient.