@theStack 2021-12-16
> I don't think it's even necessary to use `IsMine`. We already filter for addresses that are ours within the address book itself by setting `CAddressBookData.purpose` to `receive`. So you can just check for that string rather than doing `IsMine`. It will certainly be more efficient.

Thanks for the hint! In the current version of the PR though, `IsMine` is not called in the hot spot (mapWallet/CTxOut iteration loop) anymore. It may still make sense to check for the `CAddressBookData.purpose` instead of using `IsMine` in the preparation loop, which creates the "needles" of the search. Will take a deeper look on the weekend.