@kristapsk 2022-04-19
Comments by @theStack and @achow101 calmed me down, but will try to benchmark myself just in case too.