@theStack 2022-05-12
@furszy: Thanks for reviewing!

> Plus, probably something like [furszy@e8d8607](https://github.com/furszy/bitcoin-core/commit/e8d860742ca1c834c0e4a31a38bd14ab3d681f5d) would be good to include here or on a follow-up PR.
> 
> Basically, inside `GetReceived`, we can return early if the label provided by `getreceivedbylabel` is not in the address book. Otherwise, we walk through all the wallet txs + outputs for no reason (`output_scripts` is empty).

This seems to be an optimization for a very special case (passing a label that _is_ in the address book, but doesn't apply to any address that is ours), hence I think this would fit better for a follow-up PR. Would you mind opening one? If you want, can already do that now, branching off of this PR.