@furszy 2022-05-12
Hey, sure. Will do it.

> This seems to be an optimization for a very special case (passing a label that is in the address book, but doesn't apply to any address that is ours)

Actually, can happen when the user inputs a label that is not in the address book as well (aka `wallet.GetLabelAddresses(label)` returns an empty set).