@theStack 2022-05-12
> > This seems to be an optimization for a very special case (passing a label that is in the address book, but doesn't apply to any address that is ours)
> 
> Actually, can happen when the user inputs a label that is not in the address book as well (eg: `wallet.GetLabelAddresses(label)` returning an empty set).

Good point. So it would be an improvement of the user-interface, which is maybe even more important; right now if the passed label doesn't exist (due to e.g. a typo from the user), the amount of 0 is returned instead of an error.