@furszy 2022-05-12
yeah, for that I added an error there "Address/es not found in wallet" (a general error for both `getreceivedbylabel` and `getreceivedbyaddress` if the address/label is not in the wallet).

But we could customize it a bit more too; have an error string targeting `getreceivedbyaddress` -> "<address> not found in wallet" and another one for `getreceivedbylabel` -> "<label> not found in wallet".

In any case, just opened #25122. Will rebase it on master once this one gets merged.